+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * [Sigil] Introduced the `HEEX` modifier to compile Phoenix LiveView HEEx templates, supporting assigns, `{ }` expressions, and components.
   * Added utility function `MDEx.rendered_to_html/1` to convert a `%Phoenix.LiveView.Rendered{}` to static HTML string.
   * [Options] Rename `:unsafe_` to `:unsafe` in `:render` options. The old `:unsafe_` option still works.
+  * Allow to disable the built-in syntax highlighter
 
 ### Fixes
   * [Sigils] Remove `smart: true` option to avoid conversion of punctuation.